backend/requirements.txt
autopep8==1.5.4
Django==2.2
django-nested-admin==3.3.2
djangorestframework==3.12.0
Pillow==7.2.0
pycodestyle==2.6.0
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
